requirements.txt
# crucible/requirements.txt
PyQt6==6.8.1
PyQt6-WebEngine==6.8.0
requests==2.32.3
Pillow==11.1.0
